// components/forms/bridge.d.ts
export type BridgeKycFormData = {
    email: string;
    password: string;
    companyName: string;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    ssn: string;
    dob: string;
    streetAddress: string;
    streetAddress2: string;
    city: string;
    state: string;
    postalCode: string;
    country: string;
    signedAgreementId: string;
}